% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sf_moist_forest.R
\name{create_sf_moist_forest}
\alias{create_sf_moist_forest}
\title{Create sf object of moist forest according to FAO's criteria.
 MAP >=1500 mm
<5 months where precipitations < 100 mm}
\usage{
create_sf_moist_forest(EPSG = 3163, forest, monthly_precipitation)
}
\arguments{
\item{EPSG}{int. target EPSG, default 3163}

\item{forest}{sf object. output of "read_sf" of "sf" library with boolean values of forest.}

\item{monthly_precipitation}{multilayers stars object. output of "read_stars" of "stars" library with montly mean of precipitation in mm.}
}
\value{
sf object. with only moist forest area.
}
\description{
Create sf object of moist forest according to FAO's criteria.
 MAP >=1500 mm
<5 months where precipitations < 100 mm
}
