% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_JSDM.R
\name{data_JSDM}
\alias{data_JSDM}
\title{Create dataframe with values of each variable on each site}
\usage{
data_JSDM(EPSG = 3163, latlon_site, area_site, path_tiff, log_area_site = TRUE)
}
\arguments{
\item{EPSG}{int. target EPSG, default 3163}

\item{latlon_site}{float matrix with 2 columns. Columns are longitude and latitude in this order.}

\item{area_site}{float vector same dimension then number of row `latlon`. area of each site in m²}

\item{path_tiff}{character. path to Tiff file with explanatories variables}

\item{log_area_site}{boolean. Replace area of each site by log area of each site. Anyway, this variable is scale.}
}
\value{
matrix with latitude, longitude and all variables. If `latlon_site` has row names, output has same row name.
}
\description{
Create dataframe with values of each variable on each site
}
