% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JSDM_bino_pro.R
\name{JSDM_bino_pro}
\alias{JSDM_bino_pro}
\title{Run jSDM_binomial_probit from jSDM package and plot results for some species. Plots can be display and/or save.}
\usage{
JSDM_bino_pro(
  PA,
  site_data,
  n_latent = 0,
  V_beta,
  mu_beta,
  nb_species_plot = 2,
  display_plot = TRUE,
  save_plot = FALSE
)
}
\arguments{
\item{PA}{int matrix. Presence Absence for each species (col) on each site (row). Avoid empty row or column}

\item{site_data}{float matrix. Explanatories variables for each site without missing values. Same number of row than number of row in `PA`}

\item{n_latent}{int. number of latent variables to use in the model, default is 0.}

\item{V_beta}{float vector. Variances of Normal priors for the beta parameters.}

\item{mu_beta}{float vector. Means of Normal priors for the beta parameters.}

\item{nb_species_plot}{number of species whose results are display.}

\item{display_plot}{boolean. If TRUE, display all plot, default is TRUE.}

\item{save_plot}{boolean. Save in local in .png all plot in folder plot, default is FALSE.}
}
\value{
jSDM object. output of 'jSDM_binomial_probit' from 'jSDM' library
}
\description{
Run jSDM_binomial_probit from jSDM package and plot results for some species. Plots can be display and/or save.
}
