% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_interpolation.R
\name{knn_interpolation_jSDM}
\alias{knn_interpolation_jSDM}
\title{Interpole alpha and all latent variables with knn algorithm. Then save interpolation on disk in output folder.}
\usage{
knn_interpolation_jSDM(jSDM_binom_pro, k = 5, coord_site, sf_forest)
}
\arguments{
\item{jSDM_binom_pro}{object of class jSDM. output of "jSDM_binomial_probit" of "jSDM" library}

\item{k}{int. number of neighbours to consider in knn algorithm, default is 5.}

\item{coord_site}{dataframe. columns name must contain latitude and longitude as names.}

\item{sf_forest}{sf object. area where interpolation will be done.}
}
\description{
Interpole alpha and all latent variables with knn algorithm. Then save interpolation on disk in output folder.
}
