% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_min_species_group.R
\name{max_min_species_group}
\alias{max_min_species_group}
\title{Get species with estimated probabilities of presence highest for each group. Same thing with lowest probabilities.}
\usage{
max_min_species_group(
  list_theta_path,
  stars_pixels_group,
  nb_top_species = 20,
  save_csv = TRUE
)
}
\arguments{
\item{list_theta_path}{character vector. full path to Tiff files containing species probabilities of presence to considers.}

\item{stars_pixels_group}{stars object. each pixel contains values of its group number.}

\item{nb_top_species}{int. Number of species to display for each group, for lists of lowest and highest probabilities, default is 20.}

\item{save_csv}{boolean. Allows to save return in output folder in .csv file, default is TRUE.}
}
\value{
dataframe. For each group, species name with highest probabilities of presence, mean probabilities, sd probabilities. Same for species with lowest probabilities of presence.
}
\description{
Get species with estimated probabilities of presence highest for each group. Same thing with lowest probabilities.
}
