% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCA_or_tSNE_on_prop_pres_est.R
\name{PCA_or_tSNE_on_pro_pres_est}
\alias{PCA_or_tSNE_on_pro_pres_est}
\title{Run PCA or tSNE on pixels with as coordinates, probabilities of presence of species. Plot Axis 1 vs axis 2, axis 2 vs axis 3 and dendrogram. Possibility to hide plot and save it in .png.}
\usage{
PCA_or_tSNE_on_pro_pres_est(
  tSNE = TRUE,
  PCA = FALSE,
  list_theta_path,
  var_exp_stars = NULL,
  display_plot = TRUE,
  save_plot = FALSE
)
}
\arguments{
\item{tSNE}{boolean. compute tSNE reduction of dimensions algorithm. If TRUE, `PCA` must be FALSE, default is TRUE.}

\item{PCA}{boolean. compute PCA reduction of dimensions algorithm. If TRUE, `tSNE` must be FALSE, default is FALSE.}

\item{list_theta_path}{character vector.  with full path of Tiff files who contain probabilities of presence of each species to consider.}

\item{var_exp_stars}{multilayer stars object. optional, only consider if `PCA` = TRUE. Plot as supplementary variables, default is NULL.}

\item{display_plot}{boolean. Show plot, default is TRUE.}

\item{save_plot}{boolean. Save plots in plot folder with png format.}
}
\value{
float matrix. with as much row as number of pixels and 3 columns for tSNE algorithm and as columns as number of axis with more than 1% of explained variance for PCA.
}
\description{
Run PCA or tSNE on pixels with as coordinates, probabilities of presence of species. Plot Axis 1 vs axis 2, axis 2 vs axis 3 and dendrogram. Possibility to hide plot and save it in .png.
}
