% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pres_est_one_species.R
\name{plot_pres_est_one_species}
\alias{plot_pres_est_one_species}
\title{Plot presence absence  and estimated probabilities of presence for one species. Possibility to save plots}
\usage{
plot_pres_est_one_species(
  jSDM_binom_pro,
  species_to_plot = colnames(jSDM_binom_pro$model_spec$presence_data)[1],
  coord_site,
  country_name = NULL,
  latlon_output = NULL,
  display_plot = TRUE,
  save_plot = FALSE
)
}
\arguments{
\item{jSDM_binom_pro}{object of class jSDM. output of "jSDM_binomial_probit" of "jSDM" library}

\item{species_to_plot}{character. species to plot, default is the first species of the list.}

\item{coord_site}{dataframe. columns name must contain latitude and longitude as names.}

\item{country_name}{character. English name of the country where are inventory sites, default is NULL.}

\item{latlon_output}{float vector. inconsistent with `country_name` coord of output box is this format {lon_min, lat_min, lon_max, lat_max}, default is NULL.}

\item{display_plot}{boolean. show plot, default is TRUE.}

\item{save_plot}{boolean. Write plot in plot folder as .png files, default is FALSE.}
}
\description{
Plot presence absence  and estimated probabilities of presence for one species. Possibility to save plots
}
