% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prob_pres_species.R
\name{plot_prob_pres_interp}
\alias{plot_prob_pres_interp}
\title{Create plot with probabilities of presence interpolated for a choosen species. Plot can be hide and/or save in plot folder}
\usage{
plot_prob_pres_interp(
  theta_stars,
  species_to_plot = names(theta_stars)[1],
  country_name = NULL,
  country_sf = NULL,
  display_plot = TRUE,
  save_plot = FALSE
)
}
\arguments{
\item{theta_stars}{multilayer stars_object. from files save in prob_est_species_forest function.}

\item{species_to_plot}{character. name of one layer of `theta_stars`, default is first one.}

\item{country_name}{character. optional, for display border of country on map, default is NULL.}

\item{country_sf}{sf object. optional, inconsistent with `country_name`. Display borders on map, default is NULL.}

\item{display_plot}{boolean. Display plot, default is TRUE.}

\item{save_plot}{boolean. Save plot in .png format in folder plot, default is FALSE.}
}
\description{
Create plot with probabilities of presence interpolated for a choosen species. Plot can be hide and/or save in plot folder
}
