% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RGB_group_by_color.R
\name{plot_RGB_group_by_color}
\alias{plot_RGB_group_by_color}
\title{Create plots with coordinates from dimensional reduction algorithm and plot pixel group with multiple colors. Plot can be hide and/or save.}
\usage{
plot_RGB_group_by_color(
  stars_pixels_group,
  coord_pixel,
  country_name = NULL,
  country_sf = NULL,
  display_plot = TRUE,
  save_plot = FALSE
)
}
\arguments{
\item{stars_pixels_group}{stars object. with group number in each pixel usefull, NA in others pixels.}

\item{coord_pixel}{flot matrix. coordinates of each pixel, obtained with `PCA_or_tSNE_on_pro_pres_est` function.}

\item{country_name}{character. optional, add country's border to plots , default is NULL.}

\item{country_sf}{sf object. optional, add border to plots, default is NULL.}

\item{display_plot}{boolean. Show plots, default is TRUE.}

\item{save_plot}{boolean. Save plots in plot folder as .png file, default is FALSE.}
}
\description{
Create plots with coordinates from dimensional reduction algorithm and plot pixel group with multiple colors. Plot can be hide and/or save.
}
