% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_species_richness.R
\name{plot_species_richness}
\alias{plot_species_richness}
\title{Plot species richnessobserved and  estimated probabilities of presence for all sites. Possibility to save plots}
\usage{
plot_species_richness(
  jSDM_binom_pro,
  coord_site,
  country_name = NULL,
  latlon_output = NULL,
  display_plot = TRUE,
  save_plot = FALSE
)
}
\arguments{
\item{jSDM_binom_pro}{object of class jSDM. output of "jSDM_binomial_probit" of "jSDM" library}

\item{coord_site}{dataframe. columns name must contain latitude and longitude as names.}

\item{country_name}{character. English name of the country where are inventory sites, default is NULL.}

\item{latlon_output}{float vector. inconsistent with `country_name` coord of output box is this format {lon_min, lat_min, lon_max, lat_max}, default is NULL.}

\item{display_plot}{boolean. show plot, default is TRUE.}

\item{save_plot}{boolean. Write plot in plot folder as .png files, default is FALSE.}
}
\description{
Plot species richnessobserved and  estimated probabilities of presence for all sites. Possibility to save plots
}
