% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_species_richness_interpolated.R
\name{plot_species_richness_interpolated}
\alias{plot_species_richness_interpolated}
\title{Create plot with species richness with optional sf border.}
\usage{
plot_species_richness_interpolated(
  list_theta_path,
  country_name = NULL,
  country_sf = NULL,
  display_plot = TRUE,
  save_plot = FALSE,
  save_tif = FALSE
)
}
\arguments{
\item{list_theta_path}{character vector. with full path of Tiff files who contain probabilities of presence of each species to consider.}

\item{country_name}{character. optional, for display border of country on map, default is NULL.}

\item{country_sf}{sf object. optional, inconsistent with `country_name`. Display borders on map, default is NULL.}

\item{display_plot}{boolean. Display plot, default is TRUE.}

\item{save_plot}{boolean. Save plot in .png format in folder plot, default is FALSE.}

\item{save_tif}{boolean. Save species richness in .tif file in folder output, default is FALSE.}
}
\value{
terra object. species richness in designated area.
}
\description{
Create plot with species richness with optional sf border.
}
