% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_est_species_forest.R
\name{prob_est_species_forest}
\alias{prob_est_species_forest}
\title{Create tif files with probabilities of presences for each species on a given area. Results are save in output/theta}
\usage{
prob_est_species_forest(
  alpha_stars,
  latent_var_stars,
  jSDM_binom_pro,
  data_stars
)
}
\arguments{
\item{alpha_stars}{stars object. with centered values in 0.}

\item{latent_var_stars}{multilayer stars object. with as layer as latent variables. Make sure each latent variable is scale.}

\item{jSDM_binom_pro}{object of class jSDM. output of "jSDM_binomial_probit" of "jSDM" library.}

\item{data_stars}{multilayer stars object. with same explanatories variables whom in `jSDM_binom_pro`. Make sure your explanatories variable are scale.}
}
\description{
Create tif files with probabilities of presences for each species on a given area. Results are save in output/theta
}
