% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproject_sf_stars.R
\name{reproject_sf_stars}
\alias{reproject_sf_stars}
\title{Reproject shapefile with right EPSG}
\usage{
reproject_sf_stars(EPSG = 3163, sf_stars_object)
}
\arguments{
\item{EPSG}{target EPSG, default 3163}

\item{sf_stars_object}{sf or stars object to reproject}
}
\value{
same object then  `sf_stars_object` with `EPSG` projection
}
\description{
Reproject shapefile with right EPSG
}
