% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensibility.R
\name{sensitivity}
\alias{sensitivity}
\title{Process Sensitivity (species rightly predicted as present) on inventory sites.}
\usage{
sensitivity(PA, jSDM_binom_pro)
}
\arguments{
\item{PA}{dataframe. containing only 0 and 1 with colnames with species names.}

\item{jSDM_binom_pro}{object of class jSDM. output of "jSDM_binomial_probit" of "jSDM" library}
}
\value{
float vector. percentage of species rightly predicted as present on each inventory site.
}
\description{
Process Sensitivity (species rightly predicted as present) on inventory sites.
}
