% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specificity.R
\name{specificity}
\alias{specificity}
\title{Process Specificity (species rightly predicted as absent) on inventory sites.}
\usage{
specificity(PA, jSDM_binom_pro)
}
\arguments{
\item{PA}{dataframe. containing only 0 and 1 with colnames with species names.}

\item{jSDM_binom_pro}{object of class jSDM. output of "jSDM_binomial_probit" of "jSDM" library}
}
\value{
float vector. percentage of species rightly predicted as absent on each inventory site.
}
\description{
Process Specificity (species rightly predicted as absent) on inventory sites.
}
