% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sf_forest.R
\name{create_sf_forest}
\alias{create_sf_forest}
\title{Create sf object of forest with minimal altitude and percentage of forest}
\usage{
create_sf_forest(
  EPSG = 3163,
  altitude_min = 0,
  percentage_min = 50,
  elevation,
  forest
)
}
\arguments{
\item{EPSG}{int. target EPSG, default 3163}

\item{altitude_min}{float. minimal altitude to consider forest, default 0}

\item{percentage_min}{int. percentage minimal of forest to consider pixels as forest, default 50}

\item{elevation}{terra object. output of "rast" of "terra" library with elevation in meters}

\item{forest}{terra object. output of "rast" of "terra" library with percentage of forest}
}
\value{
sf object. with only forest higher than `altitude_min` & `percentage_min`
}
\description{
Create sf object of forest with minimal altitude and percentage of forest
}
